requirements.txt
# 戶政門牌爬蟲依賴套件

# HTTP 請求
requests>=2.28.0

# OCR 驗證碼識別
ddddocr>=1.4.0
Pillow>=9.0.0

# MySQL 資料庫
pymysql>=1.0.0
cryptography>=41.0.0

# 其他
python-dotenv>=1.0.0
